requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
fonttools==4.59.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
holidays==0.78
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.6.3
narwhals==1.48.0
numba==0.61.2
numpy==2.2.6
openai==0.28.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
prophet==1.1.7
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyod==2.0.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stanio==0.5.1
streamlit==1.47.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
pymongo
google-genai>=0.3.0
bcrypt>=4.0.0
plotly>=5.20.0
python-docx>=1.0.0
cryptography>=41.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
boto3>=1.28.0
gspread>=6.0.0
google-auth>=2.20.0